web/requirements.txt
Django==1.11.7
gunicorn
psycopg2
django-haystack
elasticsearch
django-configurations
django-tables2
django-bootstrap3
django-filter
dj_email_url
requests

shapely
geopy
django-leaflet
django-geojson [field]

django-notifications-hq

django-debug-toolbar

# excel
xlrd
tablib

# кэширование
redis
django-redis
django-cacheops

django-extensions


# sentry
raven

#json
django-json-widget

#selery
Celery
celery[redis]
django-celery-results
django-celery-beat
flower